# PERF: Optimize Mocks_Getters.cs