Hi @yeyopepe, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
|                           Method |         Mean |        Error |       StdDev | Rank |  Gen 0 | Allocated |
|--------------------------------- |-------------:|-------------:|-------------:|-----:|-------:|----------:|
| GetMocksFromConstructors_NoCtors |     238.3 ns |     39.57 ns |     23.55 ns |    1 | 0.1373 |     288 B |
|      GetMocksFromConstructors_01 | 394,248.6 ns | 40,861.89 ns | 27,027.61 ns |    3 | 6.8359 |  15,049 B |
|      GetMocksFromConstructors_02 |  33,894.8 ns |  4,299.88 ns |  2,844.10 ns |    2 | 5.8594 |  12,302 B |

**Benchmarks after changes:**
|                           Method |         Mean |        Error |       StdDev | Rank |  Gen 0 | Allocated |
|--------------------------------- |-------------:|-------------:|-------------:|-----:|-------:|----------:|
| GetMocksFromConstructors_NoCtors |     181.3 ns |     12.93 ns |      7.70 ns |    1 | 0.1376 |     288 B |
|      GetMocksFromConstructors_01 | 260,154.3 ns | 17,538.34 ns | 10,436.79 ns |    3 | 6.8359 |  14,906 B |
|      GetMocksFromConstructors_02 |  23,924.6 ns |    826.94 ns |    546.97 ns |    2 | 5.5847 |  11,729 B |

As you can see the allocations and benchmark durations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!